Decrease maximum value for capacitive soil sensor
Initially I did the following measurements. I measured the sensor
voltage in the air, then in the water. So, I received two extreme points
based on which all the calculations are made. In fact, it takes a lot of
time for the capacitive sensor to reach the voltage level in the air and
when it will finally reach it, I believe, that my plants will be dead.
That's why I've decided to decrease the maximum value a bit.